finally fix issue in windows (should use wb instead of w)

+++ src/ci/MEDfileCompatibility.c
@@ -55,7 +55,7 @@ MEDfileCompatibility(const char* const filename,
   /* Si le fichier n'est pas un fichier hdf retourne en indiquant hdfok=false et medok=false 
      mais pas de code d'erreur */
   
-  hid_t is_accessible = H5Fis_accessible(filename, (int)H5P_DEFAULT);
+  hid_t is_accessible = H5Fis_accessible(filename, H5P_DEFAULT);
   if  ( is_accessible < 0 ) {goto SORTIE;};
 
   /* Si l'on ne peut pas déterminer la version de la bibliothèque med, renvoie une erreur */
